zkonline/requirements.txt
altgraph==0.17
asgiref==3.3.1
astroid==2.4.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-crispy-forms==1.11.0
django-formtools==2.2
django-import-export==2.5.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.9
django-simple-captcha==0.5.13
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.10
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
login==0.0.6
MarkupPy==1.14
mccabe==0.6.1
mod-wsgi @ file:///E:/Python/Python38/mod_wsgi-4.7.1-cp38-cp38-win_amd64.whl
mysqlclient @ file:///H:/Downloads/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
odfpy==1.4.1
openpyxl==3.0.6
pefile==2019.4.18
Pillow==8.1.0
PyInstaller==3.6
pylint==2.5.2
PyMySQL==0.10.1
pypiwin32==223
pytz==2020.5
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.24.0
rope==0.18.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.1
urllib3==1.25.10
wfastcgi==3.0.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
